cilH finalisation back spring security jwt

+++ it-training/src/main/java/com/ittraining/main/security/UserDetailsImpl.java
@@ -15,7 +15,7 @@ public class UserDetailsImpl implements UserDetails {
 
 	private static final long serialVersionUID = 1L;
 
-	private Long id;
+	private Integer id;
 
 	private String name;
 
@@ -28,7 +28,7 @@ public class UserDetailsImpl implements UserDetails {
 
 	private Collection<? extends GrantedAuthority> authorities;
 
-	public UserDetailsImpl(Long id, String name, String username, String email, String password,
+	public UserDetailsImpl(Integer id, String name, String username, String email, String password,
 			Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.name = name;
@@ -46,7 +46,7 @@ public class UserDetailsImpl implements UserDetails {
 				user.getPassword(), authorities);
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
